Pausing of network objects. Some statistics work.

+++ Source/pandemonium-common.h
@@ -112,7 +112,8 @@ class pandemonium_common
       }
   }
 
-  static const qint64 maximum_database_size = 2147483648;
+  static const qint64 maximum_database_size = static_cast<qint64>
+    (2147483648);
 
  private:
   pandemonium_common(void)
